+++ R/data.R
@@ -31,8 +31,8 @@
 #'
 #' ```
 #'
-#' See **Examples** section for an easy way to recode the numerical values into
-#' [factors][base::factor].
+#' See [igo_recode_igoyear()] section for an easy way to recode the numerical
+#' values into [factors][base::factor].
 #'
 #' @format
 #' [`data.frame`][data.frame()] with
@@ -63,47 +63,29 @@
 #' datasets. *Journal of Peace Research, 57*(3), 492–503.
 #' \doi{10.1177/0022343319881175}.
 #' @examples
-#'
-#' data("igo_year_format3")
+#' data("state_year_format3")
 #'
 #' # Show a glimpse
 #' library(dplyr)
-#'
-#' igo_year_format3 %>%
-#'   select(ioname:angola) %>%
+#' names(state_year_format3)
+#' state_year_format3 %>%
+#'   select(ccode:afgec) %>%
+#'   filter(year > 1990) %>%
 #'   glimpse()
 #'
-#' # Recode numerical to factors with a small function
-#' recode_category <- function(x) {
-#'   r <- dplyr::case_when(
-#'     x == 0 ~ "No Membership",
-#'     x == 1 ~ "Full Membership", x == 2 ~ "Associate Membership",
-#'     x == 3 ~ "Observer", x == -9 ~ "Missing Data",
-#'     x == -1 ~ "IGO Not In Existence", TRUE ~ NA
-#'   )
-#'
-#'   factor(r,
-#'     levels = c(
-#'       "No Membership", "Full Membership", "Associate Membership", "Observer",
-#'       "Missing Data", "IGO Not In Existence", NA
-#'     ),
-#'     exclude = NULL
-#'   )
-#' }
-#'
-#' # Take a sample
-#'
-#' sample_igo_year <- igo_year_format3 %>%
+#' # Recode numerical to factors: with a sample
+#' sample_state_year <- state_year_format3 %>%
 #'   as_tibble() %>%
-#'   select(ioname:belgium)
+#'   select(ccode:afgec) %>%
+#'   filter(year == 1990)
 #'
-#' sample_igo_year %>% glimpse()
+#' sample_state_year %>% glimpse()
 #'
 #' # Recode
-#' sample_igo_year_recoded <- sample_igo_year %>%
-#'   mutate(across(-c(ioname:year), recode_category))
+#' sample_state_year_recoded <- sample_state_year %>%
+#'   mutate(across(-c(ccode:state), igo_recode_stateyear))
 #'
-#' sample_igo_year_recoded %>% glimpse()
+#' sample_state_year_recoded %>% glimpse()
 NULL
 
 #' @title Country membership to IGO by year
@@ -138,6 +120,8 @@ NULL
 #' knitr::kable(tb, col.names = c("**Category**", "**Numerical Value**"))
 #'
 #' ```
+#' See [igo_recode_stateyear()] section for an easy way to recode the numerical
+#' values into [factors][base::factor].
 #'
 #' @format [`data.frame`][data.frame()] with
 #' `r prettyNum(nrow(igoR::state_year_format3), big.mark=",")` rows. Relevant